requirements.txt
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
torch>=1.8.0
konlpy>=0.5.2
mecab-python3>=1.0.3
beautifulsoup4>=4.9.0
requests>=2.25.0
pyyaml>=5.4.0
schedule>=1.0.0
jupyter>=1.0.0
tqdm>=4.60.0
gensim>=4.0.0
fasttext>=0.9.2
transformers>=4.5.0
scipy>=1.6.0
networkx>=2.5.0
